backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary
dj-database-url
Pillow
python-dotenv
openpyxl==3.1.5
scipy>=1.11.4
numpy>=1.26.0
